refactor: move utils to a separate folder
