+++ packages/vue3-vite/src/util/defaults.ts
